+++ l2geth/cmd/utils/flags.go
@@ -898,12 +898,20 @@ var (
 		Usage:  "Disable transactions with 0 gas price",
 		EnvVar: "ROLLUP_ENFORCE_FEES",
 	}
+<<<<<<< HEAD
 	RollupFeeThresholdDownFlag = cli.Float64Flag{
+=======
+	RollupFeeThresholdDownFlag = cli.BoolFlag{
+>>>>>>> Crunchy tiger 0.0.1 (#201)
 		Name:   "rollup.feethresholddown",
 		Usage:  "Allow txs with fees below the current fee up to this amount, must be < 1",
 		EnvVar: "ROLLUP_FEE_THRESHOLD_DOWN",
 	}
+<<<<<<< HEAD
 	RollupFeeThresholdUpFlag = cli.Float64Flag{
+=======
+	RollupFeeThresholdUpFlag = cli.BoolFlag{
+>>>>>>> Crunchy tiger 0.0.1 (#201)
 		Name:   "rollup.feethresholdup",
 		Usage:  "Allow txs with fees above the current fee up to this amount, must be > 1",
 		EnvVar: "ROLLUP_FEE_THRESHOLD_UP",