+++ go/gas-oracle/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @eth-optimism/gas-oracle
 
+<<<<<<< HEAD
 ## 0.0.3
 
 ### Patch Changes
@@ -8,6 +9,9 @@
 
 ## 0.0.2
 
+=======
+## 0.0.2
+>>>>>>> Crunchy tiger 0.0.1 (#201)
 ### Patch Changes
 
 - ce3c353b: Initial implementation of the `gas-oracle`