+++ packages/batch-submitter/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+<<<<<<< HEAD
 ## 0.3.8
 
 ### Patch Changes
@@ -10,6 +11,8 @@
   - @eth-optimism/contracts@0.4.5
   - @eth-optimism/core-utils@0.5.1
 
+=======
+>>>>>>> Crunchy tiger 0.0.1 (#201)
 ## 0.3.7
 
 ### Patch Changes