+++ packages/message-relayer/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @eth-optimism/message-relayer
 
+<<<<<<< HEAD
 ## 0.1.8
 
 ### Patch Changes
@@ -10,6 +11,8 @@
   - @eth-optimism/contracts@0.4.5
   - @eth-optimism/core-utils@0.5.1
 
+=======
+>>>>>>> Crunchy tiger 0.0.1 (#201)
 ## 0.1.7
 
 ### Patch Changes